GUIScripts/Mouse Speed Tracker/requirements.txt
libraries used : tkinter
		 messagebox
		 PyQt5.QtGui
		 PyQt5.QtWidgets
		 sys
		 math
		 time